Sort mods by thier name in ascending order
